% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorComp2.R
\name{CorComp}
\alias{CorComp}
\title{Perform pair-wise correlation of clusters based on chosen statistics}
\usage{
CorComp(compare, dataset = NULL, stat = NULL, plot = TRUE,
  heatmap = TRUE, group_n = NULL, cor_method = "pearson",
  cl_method = "ward.D2", ...)
}
\arguments{
\item{compare}{a compare object}

\item{dataset}{The dataset that you want to depict. It can be a single character or a numeric value refers to the element of the compare object.}

\item{stat}{A character or a numeric value that refers to the statistics (elemnt of the dataset) you want ot depict.}

\item{plot}{Logical. If FALSE you have to define the group_n argumentum and a
dataframe is returned with cluster names and groups a according hierachical clustering.
This dataframe can be inputed to the GoupComp function. Otherwise you can plot a dendrogram with our withour a heatmap}

\item{heatmap}{Logical. Whether a heatmap should be ploted. Only when heatmap is FALSE you can draw rectangles around hierarchical cluster.}

\item{group_n}{Numeric. Defining the number of groups which into the clusters should be sorted}

\item{cor_method}{A character string defining the  correlation coefficient you want to use: "pearson", "kendall" or "spearman". The deafault is "pearson".}

\item{cl_method}{A character string defining the agglomeration method to be used. See hclust {stats} function for details.}

\item{...}{argument passed to the hclust function}
}
\value{
this function returns a compare object retain only just part of the original clusters
}
\description{
Perform pair-wise correlation of clusters based on chosen statistics
}
