% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutDendroComp.R
\name{CutDendroComp}
\alias{CutDendroComp}
\title{Cut a dendrogram into a predefined number of groups}
\usage{
CutDendroComp(dendro, n_groups = NULL, height = NULL, plot = FALSE,
  min_num = NULL, rect_opt, ...)
}
\arguments{
\item{dendro}{an object of class hclust}

\item{n_groups}{The dataset that you want to depict. It can be a single character or a numeric value refers to the element of the compare object.}

\item{plot}{Logical. If FALSE you have to define the group_n argumentum and a
dataframe is returned with cluster names and groups a according hierachical clustering.
This dataframe can be inputed to the GoupComp function. Otherwise you can plot a dendrogram with our withour a heatmap}

\item{min_num}{Integer. Minimum size of a group.}

\item{rect_opt}{argument passed to the rect.hclust function as a list}

\item{...}{arguments passed to the plot function}
}
\value{
this function returns a compare object retain only just part of the original clusters
}
\description{
Cut a dendrogram into a predefined number of groups
}
