% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DolloImport.R, R/DolloImport2.R
\name{DolloImport}
\alias{DolloImport}
\title{DolloImport}
\usage{
DolloImport(tree, files = NULL, partitioning = FALSE, groups = NULL,
  part_name = NULL)

DolloImport(tree, files = NULL, partitioning = FALSE, groups = NULL,
  part_name = NULL)
}
\arguments{
\item{groups}{}

\item{part_name}{}
}
\value{
this function returns a \code{list} object with...

this function returns a \code{list} object with...
}
\description{
Import data from compare analysis

Import data from compare analysis
}
\examples{




}
