% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DolloImport2.R
\name{DolloImport2}
\alias{DolloImport2}
\title{DolloImport2}
\usage{
DolloImport2(tree, files = NULL, partitioning = FALSE, groups = NULL,
  part_name = NULL)
}
\arguments{
\item{part_name}{}
}
\value{
this function returns a \code{list} object with...
}
\description{
Import data from compare analysis
}
\examples{


}
