% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungi_tree.R
\docType{data}
\name{fungi_tree}
\alias{fungi_tree}
\title{A tree of fungal species and others}
\format{An object of class \code{"phylo"}; see \code{\link[ape]{read.tree}}.}
\usage{
data(fungi_tree)
}
\description{
The tree contains 71 species mainly from the Kingdom Fungi, but from other lineages as well.
}
\examples{
data(fungi_tree)
plot.phylo(fungi_tree)
}
\references{
Kiss et al.
}
\keyword{datasets}
