% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNodes.R
\name{GetNodes}
\alias{GetNodes}
\title{Get node numbers of subclades}
\usage{
GetNodes(compare, species, tips = TRUE)
}
\arguments{
\item{compare}{a compare object}

\item{species}{Character vector. At least two species which specifiy a monophyletic clade}

\item{tips}{whether include species or not}
}
\value{
this function returns a compare object with grouped clusters
}
\description{
Get node numbers of subclades
}
