% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupComp.R
\name{GroupComp}
\alias{GroupComp}
\title{Creare groups among clusters}
\usage{
GroupComp(compare, groups = NULL, event_data = NULL)
}
\arguments{
\item{compare}{a compare object}

\item{groups}{A data.frame object with cluster names in the first column and group names in the second column}

\item{event_data}{A character or a numeric vector defining datasets to modify}
}
\value{
this function returns a compare object with grouped clusters
}
\description{
Creare groups among clusters
}
