% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HclustComp.R
\name{HclustComp}
\alias{HclustComp}
\title{Perform hierachical clustering of clusters base on their events...}
\usage{
HclustComp(compare, dataset = NULL, stat = NULL, cl_method = "ward",
  ...)
}
\arguments{
\item{compare}{a compare object}

\item{dataset}{The dataset that you want to depict. It can be a single character or a numeric value refers to the element of the compare object.}

\item{stat}{A character or a numeric value that refers to the statistics (elemnt of the dataset) you want ot depict.}

\item{cl_method}{A character string defining the agglomeration method to be used. See hclust {stats} function for details.}

\item{...}{arguments passed to the hclust.vector function}
}
\value{
this function returns a compare object retain only just part of the original clusters
}
\description{
Perform hierachical clustering of clusters base on their events...
}
