% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEvents.R
\name{PlotEvent}
\alias{PlotEvent}
\title{PlotEvent}
\usage{
PlotEvent(compare, dataset = NULL, stat = NULL, data = NULL,
  cladogram = FALSE, sizemeth = c("standard", "comparable", "log"),
  comp.size = c(0, 1000), base, size = 1, node_pch_opt, tip_pch_opt,
  node_text_opt, tip_text_opt, ...)
}
\arguments{
\item{dataset}{The dataset that you want to depict. It can be a single character or a numeric value refers to the element of the compare object.}

\item{stat}{A character or a numeric value that refers to the statistics (elemnt of the dataset) you want ot depict.}

\item{data}{A character or numeric vector defining the clusters or group of clusters you want to depict.}

\item{...}{}
}
\description{
PLot data
}
\examples{



}
