% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomAovTest.R
\name{RandomAovTest}
\alias{RandomAovTest}
\title{Perform random oneway anova test}
\usage{
RandomAovTest(compare, dataset = NULL, stat = NULL, aov_groups, s)
}
\arguments{
\item{compare}{a compare object}

\item{s}{}
}
\value{
this function returns a compare object with grouped clusters
}
\description{
Perform random oneway anova test
}
