% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SliceComp.R
\name{SliceComp}
\alias{SliceComp}
\title{Slice compare object}
\usage{
SliceComp(compare, clusters = NULL, data = "event_data",
  new_name_suf = "SLICED", overwrite = TRUE)
}
\arguments{
\item{compare}{a compare object}

\item{clusters}{a character vetor. With this vector you input cluster names you want to keep in the data.}

\item{data}{a character or a numeric value or a vector. With this parameter the required data can be sliced.
Default is the event_data, which is the original dataset imported in R.}

\item{new_name_suf}{suffix to add to the original data name(s)}

\item{overwrite}{should the original data be replaced with the sliced version?}
}
\value{
this function returns a compare object retain only just part of the original clusters
}
\description{
Slice compare object
}
