% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformComp.R
\name{TransformComp}
\alias{TransformComp}
\title{Transform compare object}
\usage{
TransformComp(compare, builtin = c("CumGain", "CumLoss", "NetGain",
  "NetCumGain"), custom = NULL)
}
\arguments{
\item{compare}{a compare object}

\item{builtin}{A character of One of the four built in transformation: cummulative gain (CumGain)
cummulative losses (CumLoss), netto gain (NetGain) and netto cummulative gain (NetCumGain).}

\item{custom}{A user defined function}
}
\value{
this function returns a compare object retain only just part of the original clusters
}
\description{
Transform compare object
}
