% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DolloImport.R
\name{DolloImport}
\alias{DolloImport}
\title{DolloImport}
\usage{
DolloImport(tree, files = NULL, path = NULL, pattern = NULL,
  partitioning = "NO", groups = NULL)
}
\arguments{
\item{groups}{}
}
\value{
this function returns a \code{list} object with...
}
\description{
Import data from compare analysis
}
