% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungi_tree_lad.R
\docType{data}
\name{fungi_tree_lad}
\alias{fungi_tree_lad}
\title{A ladderized tree of fungal species and others}
\format{An object of class \code{"phylo"}; see \code{\link[ape]{read.tree}}.}
\usage{
data(fungi_tree_lad)
}
\description{
The tree contains 71 species mainly from the Kingdom Fungi, but from other lineages as well.
This tree is the ladderized form of fungi_tree
}
\references{
Kiss et al.
}
\keyword{datasets}
