% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LadderizeTree.R
\name{LadderizeTree}
\alias{LadderizeTree}
\title{ladderize tree}
\usage{
LadderizeTree(tree, temp_file = "temp", orientation = "left")
}
\arguments{
\item{tree}{a \code{phylo} object}

\item{temp_file}{a character specifiying a a temporary directory see details}

\item{orientation}{}
}
\value{
this function returns \code{phylo} object which is ladderized with new tip label order
}
\description{
ladderizing a phylo object by not just indexing but reordering the tip labels
}
\examples{

}
