% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_event.R
\name{retrieve_event}
\alias{retrieve_event}
\title{Retrieve event}
\usage{
retrieve_event(compare, partitioning = FALSE, groups = NULL,
  part_name = NULL)
}
\arguments{
\item{part_name}{}
}
\value{
this function returns a compare object with additional event matrices
}
\description{
Retrieve event
}
